Removed several unnecessary type copies that caused bugs
The nature of the bug was never determined as the system would
segfault within a copy operation, but "non-deterministically."
That is, somehow, the place it segfaulted kept changing, and was
often within a std.write call for data determined to be valid.
In any case, the offending copies were determined to be unnecessary,
removed, and magically the bug is gone. Wooooo.