Fixed a bug where function return values only _happened_ to work
That is, the value from rax was never explicitely set to r8 after
the return of a function, so it was only that we crazy-inefficiently
use r8 for _everything_ that any function return values actually
worked. When using an extern func that actually returned a value,
everything blew up because the value from rax was never grabbed.
So that's fixed.