Fixed calls to std.algorithm.setSymmetricDifference
With the release of Dv2.068, and presumably due in part to the changes
to the internals of associative arrays in that release, the
calls to setSymmetricDifference in the Mellow compiler (which assume
sorted input) began to fail inexplicably.

Of the three calls in the compiler to setSymmetricDifference, two
of them took the .keys() of an associative array as one argument. It
is very likely the case that the pre-2.068 associative array
implementation just-happened-to yield sorted keys all this time.

However, setSymmetricDifference specifically mandates that its inputs
are sorted prior to being passed. This commit resolves that, and
we should hopefully see builds passing again.