All groundwork should now be in place. This commit has debug code
The function prologue now does everything necessary to set up for
the stack reallocation. If we uncomment "call __realloc_stack",
the program will segfault _after_ the reallocation is successfully
made and the new rsp is calculated, but we're unsure exactly where.
It may be literally the first stack access, either because of a bad
reallocation, or a bad new-rsp calculation, or it could segfault
some time later, which would be more challenging to debug.