DOESN'T PASS TESTS: Implement GC for stack-discoverable strings
Note two things:
1) If a string is tucked into a datastructure, it is not stack-discoverable,
and will be garbage collected even though it isn't garbage. This is just
the result of not having the full set of mutually-recursive marking
functions implemented, ie, this problem is intractible until GC is fully
implemented.
2) Most of the test suite fails, with segfaults. Haven't done any triage,
but, we're screwing around with memory, so. Hopefully the problems aren't
hard to track down, and hopefully the answer is simply, again, that GC must
be fully implemented for it to really work at all.