Enforce --no-pie for compilation/linking.
Some time in ~2017, PIE was enabled by default for GCC, at
least in Arch.

https://git.archlinux.org/svntogit/packages.git/commit/trunk?h=packages/gcc&id=5936710c764016ce306f9cb975056e5b7605a65b

This messes with our current assumptions for our generated
assembly, and causes issues with linking. We may or may not
want to actually _really_ fix this later, but for now, this
change allows us to actually link executables.